requirements.txt
# Core dependencies
opencv-python
numpy
insightface

# Deep learning and face recognition
tensorflow>=2.13.0
facenet-pytorch>=2.5.3
torch>=2.0.0
torchvision>=0.15.0

# Security and encryption
cryptography>=41.0.0
bcrypt>=4.0.0
argon2-cffi>=23.1.0

# CLI and utilities
click
tqdm>=4.65.0
colorama>=0.4.6

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0